+++ src/app/components/edition/edit-skills/edit-skills.component.ts
@@ -1,41 +1,40 @@
 import { Component, EventEmitter, Input, Output, OnInit, OnDestroy } from '@angular/core';
-import { UntypedFormBuilder } from '@angular/forms';
 import { AlertInput, AlertOptions } from '@ionic/angular';
 import { TranslateService } from '@ngx-translate/core';
+import { type } from 'os';
 import { Subject, takeUntil } from 'rxjs';
-import { ATRIBUTE_FIELDS } from 'src/app/constants/constants';
+import { SKILL_TYPES } from 'src/app/constants/constants';
 import { MAGIC_NUMBERS } from 'src/app/constants/number.constants';
-import { cloneAtributes } from 'src/app/controller/atribute.controller';
 import { ModificatorController } from 'src/app/controller/modificatorController';
-import { Atribute, Atributes } from 'src/app/model/atributes';
+import { cloneSkills } from 'src/app/controller/skillController';
 import { Character } from 'src/app/model/character';
 import { Modificator } from 'src/app/model/modificator';
+import { Skill } from 'src/app/model/skill';
 import { getSymbol } from 'src/app/pages/detail/detail.page.data.helper';
 import { CharactersService } from 'src/app/services/characters.service';
 import { easyConfirmAlert, openAlert } from 'src/app/utils/alert.utils';
 import { TableDataConfiguration } from '../../table/table.component';
 
 @Component({
-  selector: 'app-edit-atributes',
-  templateUrl: './edit-atributes.component.html',
-  styleUrls: ['./edit-atributes.component.scss'],
+  selector: 'app-edit-skills',
+  templateUrl: './edit-skills.component.html',
+  styleUrls: ['./edit-skills.component.scss'],
 })
-export class EditAtributesComponent implements OnInit, OnDestroy{
+export class EditSkillsComponent implements OnInit, OnDestroy{
 
   @Input() title = '';
-  @Input() list: any = [];
+  @Input() type = '';
   @Output() saveCharacter: EventEmitter<any> = new EventEmitter<any>();
   @Output() exitModal: EventEmitter<any> = new EventEmitter<any>();
 
-  public atributeFields = ATRIBUTE_FIELDS;
-  public atributes: Atributes = new Atributes();
+  public skills: Skill[] = [];
 
-  public selectedAtribute = '';
-  public selectedAtributeName = '';
+  public selectedSkill: Skill = new Skill('', 0);
   public modData: any[] = [];
   public monDataConf: TableDataConfiguration = this.getModDataConfiguration();
-  public inAtributes: Atributes = new Atributes();
+  public inSkills:  Skill[] = [];
   public character: Character = new Character('pepe');
+  public skillTypes = SKILL_TYPES;
   private ngUnsubscribe: Subject<void> = new Subject<void>();
 
   constructor(
@@ -44,7 +43,6 @@ export class EditAtributesComponent implements OnInit, OnDestroy{
     ){}
 
   ngOnInit(): void {
-
     this.charactersServce.character
       .pipe(takeUntil(this.ngUnsubscribe))
       .subscribe((character)=> {
@@ -63,8 +61,8 @@ export class EditAtributesComponent implements OnInit, OnDestroy{
   }
 
   public handleChange(event: any){
-    this.selectedAtribute = event.detail.value;
-    this.selectedAtributeName = this.translate.instant(`EDIT.ATRIBUTES.${this.selectedAtribute}`);
+    const foundSkill =  this.skills.find( skill=> skill.getName() === event.detail.value);
+    this.selectedSkill = foundSkill ? foundSkill : this.selectedSkill;
     this.setModDdata();
   }
 
@@ -72,47 +70,34 @@ export class EditAtributesComponent implements OnInit, OnDestroy{
     easyConfirmAlert(
       this.translate.instant('EDIT.MAIN_INFO.CHANGES_LOST'),
       () => {
-        this.character.setAtributes(this.inAtributes);
+        if(this.type === SKILL_TYPES.PRIMARY){
+          this.character.setPrymarySkills(this.inSkills);
+        }else if(this.type === SKILL_TYPES.SECONDARY){
+          this.character.setSecondarySkills(this.inSkills);
+        }
         this.fetch();
         this.exitModal.emit();
       },
       this.translate);
   }
 
-   public getAtribute(): Atribute {
-    const atributeGetter = {
-      [ATRIBUTE_FIELDS.AGILITY]: this.atributes.getAgility(),
-      [ATRIBUTE_FIELDS.APPEAR]: this.atributes.getAppearance(),
-      [ATRIBUTE_FIELDS.COMMUN]: this.atributes.getComunication(),
-      [ATRIBUTE_FIELDS.MIND]:  this.atributes.getMind(),
-      [ATRIBUTE_FIELDS.REFLEX]:  this.atributes.getReflexes(),
-      [ATRIBUTE_FIELDS.RESIST]: this.atributes.getResistance(),
-      [ATRIBUTE_FIELDS.SENSE]: this.atributes.getSense(),
-      [ATRIBUTE_FIELDS.SOUL]: this.atributes.getSoul(),
-      [ATRIBUTE_FIELDS.STRENGTH]: this.atributes.getStrength(),
-    };
-    return atributeGetter[this.selectedAtribute];
-  }
-
   public getModDataConfiguration(): TableDataConfiguration{
     const transLations = this.translate.instant('SHARED');
-    return {
-      columns: [
+    return { columns: [
         {id: 'name', name: transLations.DESC},
         {id: 'value', name: transLations.VALUE},
         {id: 'partial', name: transLations.PARCIAL}
-      ]
-    }
+      ]};
   }
 
-  public addAtribute() {
-    const target: number = this.getAtribute().getValue() + MAGIC_NUMBERS.N_1;
-    this.getAtribute().setValue(target > MAGIC_NUMBERS.N_6 ? this.getAtribute().getValue() : target)
+  public addSkill() {
+    const target: number = this.selectedSkill.getLevel() + MAGIC_NUMBERS.N_1;
+    this.selectedSkill.setLevel(target > MAGIC_NUMBERS.N_6 ? this.selectedSkill.getLevel() : target)
   }
 
-  public removeAtribute() {
-    const target: number = this.getAtribute().getValue() - MAGIC_NUMBERS.N_1;
-    this.getAtribute().setValue(target < MAGIC_NUMBERS.N_1 ? this.getAtribute().getValue() : target)
+  public removeSkill() {
+    const target: number = this.selectedSkill.getLevel() - MAGIC_NUMBERS.N_1;
+    this.selectedSkill.setLevel(target < MAGIC_NUMBERS.N_1 ? this.selectedSkill.getLevel() : target)
   }
 
 
@@ -122,21 +107,9 @@ export class EditAtributesComponent implements OnInit, OnDestroy{
       header: this.translate.instant('EDIT.NEW_MOD'),
       message: this.translate.instant('EDIT.ADD_MOD_TEXT'),
       inputs: [
-        {
-          type: 'text',
-          placeholder:  texts.DESC,
-          name: 'desc'
-        },
-        {
-          type: 'number',
-          placeholder:  texts.VALUE,
-          name: 'value'
-        },
-        {
-          type: 'checkbox',
-          name: 'partial',
-          label: texts.PARCIAL,
-        }
+        { type: 'text', placeholder:  texts.DESC, name: 'desc' },
+        { type: 'number', placeholder:  texts.VALUE,  name: 'value' },
+        { type: 'checkbox', name: 'partial', label: texts.PARCIAL }
       ],
       buttons: [
         { text:  texts.CANCEL, cssClass: 'alert-secondaryButton' },
@@ -145,16 +118,15 @@ export class EditAtributesComponent implements OnInit, OnDestroy{
           cssClass: 'alert-primaryButton',
           handler:  (data: any) => {
             // VALIDATE
-            this.getAtribute().getMods().push(new Modificator(data.value, data.desc, data.partial ===  'on'));
-            this.character.setAtributes(this.atributes);
-            this.setModDdata();
+            this.selectedSkill.getMods().push(new Modificator(data.value, data.desc, data.partial ===  'on'));
+            this.setChatacterSkills();
     }}]};
     openAlert(alertParams);
   }
 
   public removeMod() {
     const inputs: AlertInput[] =[];
-    this.getAtribute().getMods().forEach( mod => {
+    this.selectedSkill.getMods().forEach( mod => {
       inputs.push({ type: 'radio',  value: this.getModDetails(mod), label: this.getModDetails(mod) });
     });
 
@@ -167,11 +139,10 @@ export class EditAtributesComponent implements OnInit, OnDestroy{
           text:  this.translate.instant('SHARED.DELETE'),
           cssClass: 'alert-primaryButton',
           handler:  (data: any) => {
-            const modIndex: number =  this.getAtribute().getMods().findIndex( mod =>
+            const modIndex: number =  this.selectedSkill.getMods().findIndex( mod =>
               this.getModDetails(mod) === data );
-            this.getAtribute().getMods().splice(modIndex, 1);
-            this.character.setAtributes(this.atributes);
-            this.setModDdata();
+            this.selectedSkill.getMods().splice(modIndex, 1);
+            this.setChatacterSkills();
     }}]};
     openAlert(alertParams);
   }
@@ -180,17 +151,51 @@ export class EditAtributesComponent implements OnInit, OnDestroy{
     return ModificatorController.getModDetails(mod, this.translate);
   }
 
+  public addNewSkill(){
+    const alertParams: AlertOptions = {
+      header: this.translate.instant('EDIT.ADD_SKILL_TITLE'),
+      inputs : [{
+        type: 'text',
+        placeholder: this.translate.instant('EDIT.MAIN_INFO.NAME'),
+        name: 'skillName'
+      }],
+      buttons: [
+        { text:  this.translate.instant('SHARED.CANCEL'), cssClass: 'alert-secondaryButton' },
+        {
+          text:  this.translate.instant('EDIT.ADD'),
+          cssClass: 'alert-primaryButton',
+          handler:  (data: any) => {
+            if(this.type === SKILL_TYPES.SECONDARY && data.skillName){
+              this.character.getSecondarySkills().push(new Skill(data.skillName, MAGIC_NUMBERS.N_1));
+            }
+            this.setChatacterSkills();
+    }}]};
+    openAlert(alertParams);
+  }
+
   private fetch(){
-    this.atributes = this.character.getAtributes();
-    this.inAtributes = cloneAtributes(this.atributes);
+    if(this.type === SKILL_TYPES.PRIMARY){
+      this.skills = this.character.getPrymarySkills();
+    }else  if(this.type === SKILL_TYPES.SECONDARY){
+      this.skills = this.character.getSecondarySkills();
+    }
+    this.inSkills = cloneSkills(this.skills);
   }
 
   private setModDdata(){
-    this.modData = this.getAtribute().getMods().map( mod => ({
+    this.modData = this.selectedSkill.getMods().map( mod => ({
       name: mod.getName(),
       value: `${getSymbol( mod.getValue())}${mod.getValue()}`,
       partial: mod.isPartial() ? 'X' : '-'
     }));
   }
-}
 
+  private setChatacterSkills(){
+    if(this.type === SKILL_TYPES.PRIMARY){
+      this.character.setPrymarySkills(this.skills);
+    }else if(this.type === SKILL_TYPES.SECONDARY){
+      this.character.setSecondarySkills(this.skills);
+    }
+    this.setModDdata();
+  }
+}