+++ src/app/components/modal/edition-modal.component.ts
@@ -4,7 +4,7 @@ import { ModalController, NavParams } from '@ionic/angular';
 import { Character } from 'src/app/model/character';
 import { CharactersService } from 'src/app/services/characters.service';
 import { StorageService } from 'src/app/services/storage.service';
-import { CARD_ID, DATABASE_NAME, SECTION } from 'src/app/constants/constants';
+import { CARD_ID, DATABASE_NAME, SECTION, SKILL_TYPES } from 'src/app/constants/constants';
 
 @Component({
   selector: 'app-custom-modal',
@@ -17,6 +17,7 @@ export class EditionModalComponent implements OnInit, OnDestroy{
   public character: Character = new Character('pepe');
   public characters: Character[] = [];
   public id = CARD_ID;
+  public skillTypes = SKILL_TYPES;
   private ngUnsubscribe: Subject<void> = new Subject<void>();
 
   constructor(