+++ src/app/constants/number.constants.ts
@@ -5,5 +5,6 @@ export const MAGIC_NUMBERS = {
     N_2: 2,
     N_3: 3,
     N_6: 6,
-    N_9: 9
+    N_9: 9,
+    N_10: 10
 };