+++ src/app/pages/dices/dices.page.ts
@@ -12,7 +12,7 @@ import { getSymbol } from '../detail/detail.page.data.helper';
 })
 export class DicesPage {
 
-  public data: any = {
+  public data: {attribute: number, skill: number, modif: number} = {
     attribute: 5,
     skill: 3,
     modif: 0
@@ -20,20 +20,23 @@ export class DicesPage {
 
   constructor(private translate: TranslateService) {}
 
-  public rollDice(data: any){
+  public rollDice(diceData: any){
     const texts = this.translate.instant('DICES_PAGE.ALERT');
-    const rndInt = Math.floor(Math.random() * 10) + 1 + data.attribute + data.skill + data.modif;
-
+    const result = Math.floor(Math.random() * 10) + 1;
+    const secondResult = Math.floor(Math.random() * 10) + 1;
+    const rndInt = result + diceData.attribute + diceData.skill + diceData.modif;
+    const criticMsg = result === MAGIC_NUMBERS.N_10 ? `<p class="center bold">${texts.CRITICAL}</p><p class="dice-result center">${secondResult} </p>` : '';
+    const fumbleMsm = result === MAGIC_NUMBERS.N_1 ? `<p class="center bold">${texts.FUMBLE} </p><p class="dice-result center">${secondResult} </p>` : '';
     const alerParams: AlertOptions = {
       header: texts.HEADER,
       message: `<p class="center bold">${this.translate.instant('DICES_PAGE.ALERT.TEXT', {
-                  attribute: `${getSymbol(data.attribute)}${data.attribute}`,
-                  skill: `${getSymbol(data.skill)}${data.skill}`,
-                  modif: `${getSymbol(data.modif)}${data.modif}`
+                  attribute: `${getSymbol(diceData.attribute)}${diceData.attribute}`,
+                  skill: `${getSymbol(diceData.skill)}${diceData.skill}`,
+                  modif: `${getSymbol(diceData.modif)}${diceData.modif}`
                   })}
                 </p>
                 <p class="center">${texts.RESULT}</p>
-                <p class="dice-result center">${rndInt} </p>`,
+                <p class="dice-result center">${rndInt} </p>${criticMsg}${fumbleMsm}`,
       buttons: [{
         text: texts.AGREE,
         cssClass: 'alert-primaryButton'